# packages/schemas/recipe/openapi.yaml
...
openapi: 3.0.3
info:
  title: Commise Recipe API
  version: 1.0.0
  description: |-
    The recipe service owns this contract. It is DERIVED from the service’s authored zod
    (`src/**/*.schema.ts` plus the `@kitchensink/recipe-core` schemas those compose), never
    hand-written, and it is not a code-generation input.

    Every path is also served under a deprecated bare `/v1/...` alias for consumers configured
    outside this repo and for already-shipped mobile binaries (ADR-0011). New integrators must use
    the `/api/v1/...` spelling published here.

    Coverage is partial and the gaps are enumerated by the generator on every run: an operation with
    no response schema is one whose body has no authored zod yet, not one whose body is empty.
servers:
  - url: https://recipe.commise.app
    description: Production
  - url: https://recipe.sandbox.commise.app
    description: Sandbox
security:
  - clerkSessionToken: []
paths:
...
  /api/v1/recipes:
    post:
      operationId: createRecipe
      summary: Create a recipe owned by the caller.
      security:
        - clerkSessionToken: []
      responses:
        "201":
          description: The created recipe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        description: The recipe to create.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeRequest"
    get:
      operationId: listRecipes
      summary: List the caller's recipes, newest-updated first by default.
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: A page of the caller’s recipes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRecipes"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: page
          in: query
          required: false
          description: 1-based page number.
          schema:
            default: 1
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          description: Page size, 1..100.
          schema:
            default: 20
            type: integer
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          required: false
          description: Sort key.
          schema:
            default: updatedAt
            type: string
            enum:
              - updatedAt
              - createdAt
              - title
  /api/v1/recipes/nutrition-batch:
    post:
      operationId: getRecipeNutritionBatch
      summary: Per-serving nutrition for many recipes at once (the deferred calorie lookup).
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The nutrition state per READABLE recipe. Unreadable and unknown ids are absent from `nutrition`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeNutritionResponse"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "Answers up to 500 recipes in one call so a card grid renders immediately and fills its figures in
        afterwards. ⛔ A recipe the caller may NOT read is OMITTED from the map — absence is the authorization signal, so
        a missing key means “not for you”, never an error and never “no data”. Each present entry is a two-member
        discriminated union: `known` ALWAYS carries a number (a `0` there is a real measured zero, which water genuinely
        has) plus a `freshness` saying whether the underlying food data is current or was served from cache during an
        outage; `unaccounted` carries a `reason` and NO figure at all, so an outage can never be rendered as zero
        calories. There is deliberately no `pending` member — a state a server can emit is a spinner a server can pin
        forever.  ⚠️ POST for a READ, unlike the food service’s cacheable `GET /api/v1/foods/nutrition`: this response
        varies by caller by construction, so it must never be cached at a URL-keyed edge — and at the id cap the query
        string would exceed CloudFront’s URL limit anyway."
      requestBody:
        description: The recipes to report on.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeNutritionRequest"
  /api/v1/recipes/{id}:
    get:
      operationId: getRecipeById
      summary: Read one recipe (the owner’s, or any public recipe).
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The recipe, with photos and nutrition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such recipe, or it is not visible to the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: The recipe was erased (tombstoned).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The recipe id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
    patch:
      operationId: updateRecipe
      summary: Update the caller’s recipe (optimistic concurrency on `currentVersion`).
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The updated recipe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such recipe, or it is not visible to the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Version conflict — the recipe changed since the caller read it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The recipe id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      requestBody:
        description: The fields to change.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeRequest"
    delete:
      operationId: deleteRecipe
      summary: Soft-delete the caller’s recipe.
      security:
        - clerkSessionToken: []
      responses:
        "204":
          description: Deleted.
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such recipe, or it is not visible to the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The recipe id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
  /api/v1/recipes/{id}/clone:
    post:
      operationId: cloneRecipe
      summary: Clone a visible recipe into the caller’s collection.
      security:
        - clerkSessionToken: []
      responses:
        "201":
          description: The caller’s new clone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such recipe, or it is not visible to the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The recipe to clone.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
...
components:
  securitySchemes:
    clerkSessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Clerk session token. Verified networklessly against the instance JWT key, with the signed `azp` checked
        against the service’s authorized-party pattern.
    serviceErasureToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A signed service-principal token, used only by the internal erasure route.
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
      required:
        - code
        - message
      additionalProperties: {}
...
    PaginatedRecipes:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Recipe"
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 0
          exclusiveMinimum: true
        pageSize:
          type: integer
          minimum: 0
          exclusiveMinimum: true
        hasMore:
          type: boolean
      required:
        - data
        - total
        - page
        - pageSize
        - hasMore
...
    RecipeNutritionRequest:
      type: object
      properties:
        recipeIds:
          minItems: 1
          maxItems: 500
          type: array
          items:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - recipeIds
      additionalProperties: false
...
    RecipeNutritionResponse:
      type: object
      properties:
        nutrition:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RecipeNutritionState"
      required:
        - nutrition
      additionalProperties: false
...